TrenKayit
It is an HTTP API that can pull data from Microsoft SQL database and respond to HTTP Post and HTTP Get requests. This application records the person's name, city, boarding date and ticket price on the train to the SQL platform and brings the person information to the API by entering the personnel ID information.